+++ spec/dummy/app/models/user.rb
@@ -1,0 +1,4 @@
+class User < LHS::Record
+  endpoint ':datastore/v2/users'
+  endpoint ':datastore/v2/users/:id'
+end
