+++ spec/dummy/config/initializers/lhc.rb
@@ -1,0 +1,3 @@
+LHC.configure do |config|
+  config.placeholder(:datastore, 'http://datastore')
+end
