--- spec/support/cleanup_endpoints.rb
@@ -1,5 +1,0 @@
-RSpec.configure do |config|
-  config.before(:each) do
-    LHS::Record::Endpoints.all = {}
-  end
-end
